@@ -1,21 +1,22 @@
 name: ninja
 description: Cryptography dart library to encrypt, decrypt, sign and verify messages
-version: 3.0.1
+version: 3.0.2
 homepage: https://github.com/ninja-dart/ninja
-author: Ravi Teja Gudapati <tejainece@gmail.com>
+
+documentation:
 
 environment:
-  sdk: '>=2.3.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  crypto: ^2.1.4
-  ninja_prime: ^1.0.0
-  ninja_pem: ^1.0.0
-  ninja_asn1: ^1.0.0
-  ninja_hex: ^1.0.0
-  collection: ^1.14.12
+  crypto: ^3.0.1
+  ninja_prime: ^2.0.0
+  ninja_pem: ^2.0.0
+  ninja_asn1: ^2.0.0
+  ninja_hex: ^2.0.0
+  collection: ^1.15.0
 
 dev_dependencies:
-  test: ^1.3.0
+  test: ^1.17.3
   pedantic:
   ninja_openssl: ^1.0.4
