@@ -6,7 +6,7 @@ import 'padder.dart';
 class EmePkcs1v15Encoder implements Padder, IndividualBlockPadder {
   final Random rand;
 
-  EmePkcs1v15Encoder({Random rand}) : rand = rand ?? Random.secure();
+  EmePkcs1v15Encoder({Random? rand}) : rand = rand ?? Random.secure();
 
   /// Pads a single block
   void padBlock(int blockSize, Iterable<int> block, ByteData output) {
